requirement.txt
amqp==1.4.9
anyjson==0.3.3
arrow==0.12.1
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
Babel==2.5.3
billiard==3.3.0.23
captcha==0.2.2
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==2.1.3
cssselect==1.0.1
cycler==0.10.0
decorator==4.0.6
diff-match-patch==20121119
Django==1.9
django-appconf==1.0.2
django-cors-headers==2.2.0
django-crispy-forms==1.7.0
django-crontab==0.7.1
django-extensions==2.1.2
django-filter==1.1.0
django-filters==0.2.1
django-formtools==1.0
django-guardian==1.4.9
django-image-cropping==1.2.0
django-import-export==0.7.0
django-mdeditor==0.1.7
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==2.0.13
django-simple-captcha==0.5.6
django-suit==0.2.25
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
drfdocs==0.0.11
easy-thumbnails==2.5
echarts-python==0.1.3
entrypoints==0.2.3
et-xmlfile==1.0.1
fake-useragent==0.1.8
feedparser==5.1.3
flower==0.9.2
future==0.15.2
glob2==0.5
guacamole==0.9.2
html5lib==0.999999999
httplib2==0.11.3
hyperlink==18.0.0
idna==2.5
incremental==17.5.0
ipython-genutils==0.2.0
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
kombu==4.2.1
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.13
Naked==0.1.31
odfpy==1.3.6
olefile==0.45.1
openpyxl==2.5.4
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==4.1.1
ptyprocess==0.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycparser==2.18
pycrypto==2.6.1
pycryptodomex==3.6.5
Pygments==2.2.0
PyJWT==1.6.4
PyMySQL==0.9.2
pyOpenSSL==17.3.0
pyquery==1.2.17
python-alipay-sdk==1.7.1
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.13
qiniu==7.2.0
qrcode==6.0
raven==6.9.0
redis==2.10.6
requests==2.18.1
requests-file==1.4.2
requests-ftp==0.3.1
service-identity==17.0.0
shellescape==3.4.1
simplegeneric==0.8.1
six==1.11.0
tablib==0.12.1
tornado==5.0.1
traitlets==4.3.2
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.21.1
uWSGI==2.0.17.1
vine==1.1.4
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
wincertstore==0.2
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.4.3
